commit before deploying

+++ src/app/api/menu-items/route.js
@@ -1,17 +1,24 @@
 import mongoose from "mongoose";
 import { MenuItem } from "../../models/MenuItem";
+import { isAdmin } from "../auth/[...nextauth]/route";
 
 export async function POST(req) {
   mongoose.connect(process.env.MONGO_URL);
   const data = await req.json();
-  const menuItemDoc = await MenuItem.create(data);
-  return Response.json(menuItemDoc);
+  if (await isAdmin()) {
+    const menuItemDoc = await MenuItem.create(data);
+    return Response.json(menuItemDoc);
+  } else {
+    return Response.json({});
+  }
 }
 
 export async function PUT(req) {
   mongoose.connect(process.env.MONGO_URL);
-  const { _id, ...data } = await req.json();
-  await MenuItem.findByIdAndUpdate(_id, data);
+  if (await isAdmin()) {
+    const { _id, ...data } = await req.json();
+    await MenuItem.findByIdAndUpdate(_id, data);
+  }
   return Response.json(true);
 }
 
@@ -20,10 +27,12 @@ export async function GET(req) {
   return Response.json(await MenuItem.find());
 }
 
-export async function DELETE(req){
+export async function DELETE(req) {
   mongoose.connect(process.env.MONGO_URL);
   const url = new URL(req.url);
-  const _id = url.searchParams.get('_id')
-  await MenuItem.deleteOne({_id})
+  const _id = url.searchParams.get("_id");
+  if (await isAdmin()) {
+    await MenuItem.deleteOne({ _id });
+  }
   return Response.json(true);
 }
